add cancel plugin install, add active plugin tasks ceck

+++ src/NiceHashMinerLegacy/Plugin/MinerPluginsManager.cs
@@ -233,16 +233,20 @@ namespace NiceHashMiner.Plugin
                 // download plugin dll
                 var pluginPackageDownloaded = Path.Combine(installingPluginPath, "plugin.zip");
                 var downloadPluginOK = await DownloadFile(plugin.PluginPackageURL, pluginPackageDownloaded, downloadPluginProgressChangedEventHandler, stop);
+                if (!downloadPluginOK || stop.IsCancellationRequested) return;
                 // unzip 
                 var unzipPluginOK = await UnzipFile(pluginPackageDownloaded, installingPluginPath, zipProgressPluginChangedEventHandler, stop);
+                if (!unzipPluginOK || stop.IsCancellationRequested) return;
                 File.Delete(pluginPackageDownloaded);
 
                 // download plugin dll
                 var binsPackageDownloaded = Path.Combine(installingPluginPath, "miner_bins.zip");
                 var downloadMinerBinsOK = await DownloadFile(plugin.MinerPackageURL, binsPackageDownloaded, downloadMinerProgressChangedEventHandler, stop);
+                if (!downloadMinerBinsOK || stop.IsCancellationRequested) return;
                 // unzip 
                 var binsUnzipPath = Path.Combine(installingPluginPath, "bins");
                 var unzipMinerBinsOK = await UnzipFile(binsPackageDownloaded, binsUnzipPath, zipProgressMinerChangedEventHandler, stop);
+                if (!unzipMinerBinsOK || stop.IsCancellationRequested) return;
                 File.Delete(binsPackageDownloaded);
